+++ sdk/core/azure-core-http-netty/src/main/java/com/azure/core/http/netty/NettyAsyncHttpClientBuilder.java
@@ -18,6 +18,7 @@ import io.netty.resolver.DefaultAddressResolverGroup;
 import io.netty.resolver.NoopAddressResolverGroup;
 import reactor.netty.NettyPipeline;
 import reactor.netty.http.client.HttpClient;
+import reactor.netty.http.client.HttpResponseDecoderSpec;
 import reactor.netty.resources.ConnectionProvider;
 import reactor.netty.transport.AddressUtils;
 import reactor.netty.transport.ProxyProvider;
@@ -142,13 +143,17 @@ public class NettyAsyncHttpClientBuilder {
             addressResolverWasSetByBuilder = true;
         }
 
+        // Get the initial HttpResponseDecoderSpec and update it.
+        // .httpResponseDecoder passes a new HttpResponseDecoderSpec and any existing configuration should be updated
+        // instead of overwritten.
+        HttpResponseDecoderSpec initialSpec = nettyHttpClient.configuration().decoder();
         nettyHttpClient = nettyHttpClient
             .port(port)
             .wiretap(enableWiretap)
             .option(ChannelOption.CONNECT_TIMEOUT_MILLIS, (int) getTimeoutMillis(connectTimeout,
                 DEFAULT_CONNECT_TIMEOUT))
             // TODO (alzimmer): What does validating HTTP response headers get us?
-            .httpResponseDecoder(httpResponseDecoderSpec -> httpResponseDecoderSpec.validateHeaders(false));
+            .httpResponseDecoder(httpResponseDecoderSpec -> initialSpec.validateHeaders(false));
 
         Configuration buildConfiguration = (configuration == null)
             ? Configuration.getGlobalConfiguration()