+++ sdk/eventgrid/azure-messaging-eventgrid/src/main/java/com/azure/messaging/eventgrid/systemevents/ContainerRegistryImageDeletedEventData.java
@@ -4,11 +4,11 @@
 
 package com.azure.messaging.eventgrid.systemevents;
 
-import com.azure.core.annotation.Immutable;
+import com.azure.core.annotation.Fluent;
 import java.time.OffsetDateTime;
 
 /** Schema of the Data property of an EventGridEvent for a Microsoft.ContainerRegistry.ImageDeleted event. */
-@Immutable
+@Fluent
 public final class ContainerRegistryImageDeletedEventData extends ContainerRegistryEventData {
     /** {@inheritDoc} */
     @Override