+++ sdk/keyvault/azure-security-keyvault-administration/src/main/java/com/azure/security/keyvault/administration/KeyVaultBackupAsyncClient.java
@@ -19,11 +19,9 @@ import com.azure.core.util.polling.PollingContext;
 import com.azure.security.keyvault.administration.implementation.KeyVaultBackupClientImpl;
 import com.azure.security.keyvault.administration.implementation.KeyVaultBackupClientImplBuilder;
 import com.azure.security.keyvault.administration.implementation.KeyVaultErrorCodeStrings;
-import com.azure.security.keyvault.administration.implementation.models.FullBackupOperation;
 import com.azure.security.keyvault.administration.implementation.models.RestoreOperation;
 import com.azure.security.keyvault.administration.implementation.models.RestoreOperationParameters;
 import com.azure.security.keyvault.administration.implementation.models.SASTokenParameter;
-import com.azure.security.keyvault.administration.implementation.models.SelectiveKeyRestoreOperation;
 import com.azure.security.keyvault.administration.implementation.models.SelectiveKeyRestoreOperationParameters;
 import com.azure.security.keyvault.administration.models.KeyVaultBackupOperation;
 import com.azure.security.keyvault.administration.models.KeyVaultAdministrationException;
@@ -37,9 +35,6 @@ import reactor.core.publisher.Mono;
 
 import java.net.URL;
 import java.time.Duration;
-import java.time.Instant;
-import java.time.OffsetDateTime;
-import java.time.ZoneOffset;
 import java.util.Locale;
 import java.util.Objects;
 import java.util.function.Function;
@@ -47,6 +42,9 @@ import java.util.function.Function;
 import static com.azure.core.util.FluxUtil.monoError;
 import static com.azure.core.util.FluxUtil.withContext;
 import static com.azure.core.util.tracing.Tracer.AZ_TRACING_NAMESPACE_KEY;
+import static com.azure.security.keyvault.administration.KeyVaultAdministrationUtil.transformToLongRunningOperation;
+import static com.azure.security.keyvault.administration.KeyVaultAdministrationUtil.toLongRunningOperationStatus;
+import static com.azure.security.keyvault.administration.KeyVaultAdministrationUtil.longToOffsetDateTime;
 import static com.azure.security.keyvault.administration.implementation.KeyVaultAdministrationUtils.createKeyVaultErrorFromError;
 
 /**
@@ -306,20 +304,6 @@ public final class KeyVaultBackupAsyncClient {
             toLongRunningOperationStatus(operationStatus.toLowerCase(Locale.US)), response.getValue()));
     }
 
-    private static LongRunningOperationStatus toLongRunningOperationStatus(String operationStatus) {
-        switch (operationStatus) {
-            case "inprogress":
-                return LongRunningOperationStatus.IN_PROGRESS;
-            case "succeeded":
-                return LongRunningOperationStatus.SUCCESSFULLY_COMPLETED;
-            case "failed":
-                return LongRunningOperationStatus.FAILED;
-            default:
-                // Should not reach here
-                return LongRunningOperationStatus.fromString("POLLING_FAILED", true);
-        }
-    }
-
     /**
      * Initiates a full restore of the Key Vault.
      *
@@ -471,7 +455,7 @@ public final class KeyVaultBackupAsyncClient {
         };
     }
 
-    private static Mono<PollResponse<KeyVaultRestoreOperation>> processRestoreOperationResponse(
+    static Mono<PollResponse<KeyVaultRestoreOperation>> processRestoreOperationResponse(
         Response<KeyVaultRestoreOperation> response) {
 
         String operationStatus = response.getValue().getStatus().toLowerCase(Locale.US);
@@ -656,47 +640,12 @@ public final class KeyVaultBackupAsyncClient {
             toLongRunningOperationStatus(operationStatus.toLowerCase(Locale.US)), response.getValue()));
     }
 
-    private static <O> KeyVaultLongRunningOperation transformToLongRunningOperation(O operation) {
-        if (operation instanceof RestoreOperation) {
-            RestoreOperation restoreOperation = (RestoreOperation) operation;
-
-            return new KeyVaultRestoreOperation(restoreOperation.getStatus(), restoreOperation.getStatusDetails(),
-                createKeyVaultErrorFromError(restoreOperation.getError()), restoreOperation.getJobId(),
-                longToOffsetDateTime(restoreOperation.getStartTime()),
-                longToOffsetDateTime(restoreOperation.getEndTime()));
-        } else if (operation instanceof SelectiveKeyRestoreOperation) {
-            SelectiveKeyRestoreOperation selectiveKeyRestoreOperation = (SelectiveKeyRestoreOperation) operation;
-
-            return new KeyVaultSelectiveKeyRestoreOperation(selectiveKeyRestoreOperation.getStatus(),
-                selectiveKeyRestoreOperation.getStatusDetails(),
-                createKeyVaultErrorFromError(selectiveKeyRestoreOperation.getError()),
-                selectiveKeyRestoreOperation.getJobId(),
-                longToOffsetDateTime(selectiveKeyRestoreOperation.getStartTime()),
-                longToOffsetDateTime(selectiveKeyRestoreOperation.getEndTime()));
-        } else if (operation instanceof FullBackupOperation) {
-            FullBackupOperation fullBackupOperation = (FullBackupOperation) operation;
-
-            return new KeyVaultBackupOperation(fullBackupOperation.getStatus(), fullBackupOperation.getStatusDetails(),
-                createKeyVaultErrorFromError(fullBackupOperation.getError()), fullBackupOperation.getJobId(),
-                longToOffsetDateTime(fullBackupOperation.getStartTime()),
-                longToOffsetDateTime(fullBackupOperation.getEndTime()),
-                fullBackupOperation.getAzureStorageBlobContainerUri());
-        } else {
-            throw new UnsupportedOperationException();
-        }
-    }
-
-    private static KeyVaultLongRunningOperation restoreOperationToSelectiveKeyRestoreOperation(RestoreOperation operation) {
+    static KeyVaultLongRunningOperation restoreOperationToSelectiveKeyRestoreOperation(RestoreOperation operation) {
         return new KeyVaultSelectiveKeyRestoreOperation(operation.getStatus(),
             operation.getStatusDetails(),
             createKeyVaultErrorFromError(operation.getError()),
             operation.getJobId(),
             longToOffsetDateTime(operation.getStartTime()),
             longToOffsetDateTime(operation.getEndTime()));
-    }
-
-    private static OffsetDateTime longToOffsetDateTime(Long epochInSeconds) {
-        return epochInSeconds == null ? null
-            : OffsetDateTime.ofInstant(Instant.ofEpochSecond(epochInSeconds), ZoneOffset.UTC);
     }
 }