+++ sdk/machinelearning/azure-resourcemanager-machinelearning/src/main/java/com/azure/resourcemanager/machinelearning/models/ContainerResourceRequirements.java
@@ -22,6 +22,10 @@ public final class ContainerResourceRequirements {
     @JsonProperty(value = "containerResourceRequests")
     private ContainerResourceSettings containerResourceRequests;
 
+    /** Creates an instance of ContainerResourceRequirements class. */
+    public ContainerResourceRequirements() {
+    }
+
     /**
      * Get the containerResourceLimits property: Container resource limit info:.
      *