+++ sdk/machinelearning/azure-resourcemanager-machinelearning/src/main/java/com/azure/resourcemanager/machinelearning/models/QuotaUnit.java
@@ -8,7 +8,7 @@ import com.azure.core.util.ExpandableStringEnum;
 import com.fasterxml.jackson.annotation.JsonCreator;
 import java.util.Collection;
 
-/** Defines values for QuotaUnit. */
+/** An enum describing the unit of quota measurement. */
 public final class QuotaUnit extends ExpandableStringEnum<QuotaUnit> {
     /** Static value Count for QuotaUnit. */
     public static final QuotaUnit COUNT = fromString("Count");